gateway/requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.30.6
pydantic[email]==2.9.2
email-validator==2.2.0
pydantic-settings==2.5.2
python-jose[cryptography]==3.3.0
httpx==0.27.2
pytest==8.3.3
pytest-asyncio==0.24.0
respx==0.21.1
